# docs/developer/api.yaml
...
openapi: 3.1.0
info:
  title: OpenTalk Controller Frontend API
  description: Specifies the endpoints and structure of the OpenTalk Controller Frontend API
  contact:
    name: OpenTalk Team
    email: mail@opentalk.eu
  license:
    name: EUPL-1.2
    identifier: EUPL-1.2
  version: 0.35.0-alpha
servers:
  - url: /v1
paths:
...
  /rooms:
    get:
      tags:
        - "api::v1::rooms"
      summary: Get a list of rooms accessible by the requesting user
      description: |-
        All rooms accessible to the requesting user are returned in a list. If no
        pagination query is added, the default page size is used.
      operationId: get_rooms
      parameters:
        - name: per_page
          in: query
          description: The number of entries per page
          required: false
          schema:
            $ref: "#/components/schemas/PageSize"
        - name: page
          in: query
          description: The number of the page
          required: false
          schema:
            $ref: "#/components/schemas/Page"
      responses:
        "200":
          description: List of accessible rooms successfully returned
          headers:
            link:
              schema:
                $ref: "#/components/schemas/PageLink"
              description: Links for paging through the results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRoomsResponseBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
    post:
      tags:
        - "api::v1::rooms"
      summary: Create a new room
      description: Creates a new room with the settings given in the request body.
      operationId: post_rooms
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostRoomsRequestBody"
        required: true
      responses:
        "201":
          description: Room successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResource"
        "400":
          description: Wrong syntax or bad values such as invalid owner id received in the body
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Invalid body contents received
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
...
  "/rooms/{room_id_or_alias}/invites/{invite_code}":
    get:
      tags:
        - "api::v1::invites"
      summary: Get a room invite
      description: Returns the room invite resource
      operationId: get_invite
      parameters:
        - name: room_id_or_alias
          in: path
          description: The room id or alias for the invite
          required: true
          schema:
            $ref: "#/components/schemas/RoomIdOrAlias"
        - name: invite_code
          in: path
          description: The invite code id
          required: true
          schema:
            $ref: "#/components/schemas/InviteCode"
      responses:
        "200":
          description: Successfully retrieved the room invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
    put:
      tags:
        - "api::v1::invites"
      summary: Update an invite code
      description: Updates the field values as set in the request body.
      operationId: update_invite
      parameters:
        - name: room_id_or_alias
          in: path
          description: The room id or alias for the invite
          required: true
          schema:
            $ref: "#/components/schemas/RoomIdOrAlias"
        - name: invite_code
          in: path
          description: The invite code id
          required: true
          schema:
            $ref: "#/components/schemas/InviteCode"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutInviteRequestBody"
        required: true
      responses:
        "200":
          description: Successfully updated the room invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
    delete:
      tags:
        - "api::v1::invites"
      summary: Delete an invite code
      description: The invite code will no longer be usable once it is deleted.
      operationId: delete_invite
      parameters:
        - name: room_id_or_alias
          in: path
          description: The room id or alias for the invite
          required: true
          schema:
            $ref: "#/components/schemas/RoomIdOrAlias"
        - name: invite_code
          in: path
          description: The invite code id
          required: true
          schema:
            $ref: "#/components/schemas/InviteCode"
      responses:
        "204":
          description: Successfully deleted the room invite
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
  "/rooms/{room_id_or_alias}/sip":
    get:
      tags:
        - "api::v1::sip_configs"
      summary: Get the sip config for the specified room.
      description: |-
        Returns the sip config if available for the room, otherwise `404 NOT_FOUND`
        is returned.
      operationId: get_room_sip
      parameters:
        - name: room_id_or_alias
          in: path
          description: Either the id or the alias of the room
          required: true
          schema:
            $ref: "#/components/schemas/RoomIdOrAlias"
      responses:
        "200":
          description: The SIP config is successfully returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SipConfigResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
    put:
      tags:
        - "api::v1::sip_configs"
      summary: |-
        Modify the sip configuration of a room. A new sip configuration is created
        if none was set before.
      description: Returns the new modified sip configuration.
      operationId: put
      parameters:
        - name: room_id_or_alias
          in: path
          description: Either the id or the alias of the room
          required: true
          schema:
            $ref: "#/components/schemas/RoomIdOrAlias"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutSipConfigRequestBody"
        required: true
      responses:
        "200":
          description: The SIP configuration was updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SipConfigResource"
        "201":
          description: A new SIP configuration was created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SipConfigResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
    delete:
      tags:
        - "api::v1::sip_configs"
      summary: Delete the SIP configuration of a room.
      description: This removes the dial-in functionality from the room.
      operationId: delete_room_sip
      parameters:
        - name: room_id_or_alias
          in: path
          description: Either the id or the alias of the room
          required: true
          schema:
            $ref: "#/components/schemas/RoomId"
      responses:
        "204":
          description: The SIP configuration was successfully deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    CallInId:
      $ref: "#/components/schemas/NumericId"
      description: The id of a call-in participation
...
    CallInPassword:
      $ref: "#/components/schemas/NumericId"
      description: The password for authenticating call-in participation
...
    GetRoomsResponseBody:
      type: array
      items:
        $ref: "#/components/schemas/RoomResource"
      description: "The JSON body returned by the `/rooms` `GET` endpoint"
...
    GuestAccess:
      type: string
      enum:
        - disabled
        - waiting_room
        - direct_access
      example: waiting_room
...
    InviteCode:
      type: string
      format: uuid
      description: An invite code
      example: 00000000-0000-0000-0000-0000deadbeef
    InviteResource:
      type: object
      description: A room invite code with associated information.
      required:
        - invite_code
        - created
        - created_by
        - updated
        - updated_by
        - room_id
        - active
      properties:
        active:
          type: boolean
          description: If the invite is active
        created:
          type: string
          format: date-time
          description: The timestamp this invite was created at
        created_by:
          $ref: "#/components/schemas/PublicUserProfile"
          description: The user who created the invite
        expiration:
          type:
            - string
            - "null"
          format: date-time
          description: Optional expiration date of the invite
        invite_code:
          $ref: "#/components/schemas/InviteCode"
          description: The invite code id
        room_id:
          $ref: "#/components/schemas/RoomId"
          description: The room id for the invite
        updated:
          type: string
          format: date-time
          description: The timestamp this invite was updated at
        updated_by:
          $ref: "#/components/schemas/PublicUserProfile"
          description: The user who updated the invite
      example:
        active: true
        created: "2024-06-18T11:22:33Z"
        created_by:
          avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
          display_name: Alice Adams
          email: alice@example.com
          firstname: Alice
          id: 00000000-0000-0000-0000-0000000a11c3
          lastname: Adams
          title: ""
        expiration: ~
        invite_code: 00000000-0000-0000-0000-0000deadbeef
        room_id: 00000000-0000-0000-0000-0000abadcafe
        updated: "2024-06-20T14:16:19Z"
        updated_by:
          avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
          display_name: Alice Adams
          email: alice@example.com
          firstname: Alice
          id: 00000000-0000-0000-0000-0000000a11c3
          lastname: Adams
          title: ""
...
    NumericId:
      type: string
      description: A string containing number characters
      examples:
        - "0000000000"
      maxLength: 10
      minLength: 10
      pattern: "[0-9]+"
...
    Page:
      type: integer
      format: int64
      description: |-
        A page number used for paging in requests.

        This is `1`-based, so a `Page` with a value of `0` can not be created.
      example: 5
    PageLink:
      type: string
      description: Headers for referencing to related pages.
      example: "<https://api.example.org/resource?page=2>; rel='next', <https://api.example.org/resource?page=5>; rel='last'"
    PageSize:
      type: integer
      format: int64
      description: A page number used for paging in requests.
      example: 30
...
    PostRoomsRequestBody:
      type: object
      description: API request parameters to create a new room
      properties:
        e2e_encryption:
          type: boolean
          description: Enable/Disable e2e encryption for this room; defaults to false when not set
        enable_sip:
          type: boolean
          description: Enable/Disable sip for this room; defaults to false when not set
        guest_access:
          $ref: "#/components/schemas/GuestAccess"
          description: Guest access mode
        name:
          $ref: "#/components/schemas/RoomName"
          description: |-
            A custom name that can be used as part of the URL alias to access the room instead of the [`RoomId`](opentalk_types_common::rooms::RoomId).

            The name is only one part of the alias. The full alias additionally contains a code that is generated by the
            controller. Whether such a code is generated depends on the controller configuration.
        password:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/RoomPassword"
              description: "The password to the room, if any"
        waiting_room:
          type: boolean
          description: |-
            Indicates whether the meeting room should have the waiting room enabled.
            When not present, the waiting room will be disabled.
      example:
        e2e_encryption: false
        enable_sip: true
        guest_access: waiting_room
        name: personal-room-name
        password: ~
        waiting_room: true
...
    PublicUserProfile:
      allOf:
        - $ref: "#/components/schemas/UserInfo"
          description: General information about the user
        - type: object
          required:
            - id
            - email
          properties:
            email:
              type: string
              description: The email of the user
            id:
              $ref: "#/components/schemas/UserId"
              description: The user id
      description: |-
        Public user details.

        Contains general "public" information about a user. Is accessible to all other users.
      example:
        avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
        display_name: Alice Adams
        email: alice@example.com
        firstname: Alice
        id: 00000000-0000-0000-0000-0000000a11c3
        lastname: Adams
        title: ""
    PutInviteRequestBody:
      type: object
      description: "Body for *PUT /rooms/{room_id}/invites/{invite_code}*"
      properties:
        expiration:
          type:
            - string
            - "null"
          format: date-time
          description: Optional expiration date of the invite
      example:
        expiration: "2024-06-20T14:16:19Z"
    PutSipConfigRequestBody:
      type: object
      description: "Body for the `PUT /rooms/{room_id}/sip` endpoint"
      properties:
        lobby:
          type: boolean
          description: |-
            Enable or disable the lobby for users that join throughh SIP. Defaults
            to [`false`] when not explicity set on creation.
        password:
          $ref: "#/components/schemas/CallInPassword"
          description: |-
            Numeric code required for entering the room. If not set explicitly on
            creation, this will be set to a randomly generated number.
      example:
        lobby: true
        password: "9876543210"
...
    RoomAlias:
      type: string
      description: A human readable alias for a room.
      examples:
        - personal-room-name_0000000000000000
    RoomId:
      type: string
      format: uuid
      description: The id of a room
      example: 00000000-0000-0000-0000-0000abadcafe
    RoomIdOrAlias:
      oneOf:
        - $ref: "#/components/schemas/RoomId"
          description: A room id.
        - $ref: "#/components/schemas/RoomAlias"
          description: A room alias.
      description: An enum that represents either a room id or a room alias.
      example: 00000000-0000-0000-0000-0000abadcafe
    RoomName:
      type: string
      description: |-
        A valid name for a room.

        A room name must

        * be at least 3 characters long
        * be at most 40 characters long
        * contain only `a-z`, `0-9`, `.` and `-`
        * not be a valid UUID

        The "not a valid UUID" rule exists so that a [`RoomName`] can never be confused with a
        [`RoomId`](super::RoomId): both can appear in the same URL position and are resolved via the untagged
        [`RoomIdOrAlias`](super::RoomIdOrAlias) enum. Without it, users could hijack rooms by creating a name that
        collides with an existing room's ID.
      example: personal-room-name
    RoomPassword:
      type: string
      description: A room password
      examples:
        - v3rys3cr3t
      maxLength: 255
      minLength: 1
    RoomResource:
      type: object
      description: |-
        A Room

        Contains all room information. Is only be accessible to the owner and users with
        appropriate permissions.
      required:
        - id
        - created_by
        - created_at
        - waiting_room
      properties:
        alias:
          $ref: "#/components/schemas/RoomAlias"
          description: An optional alias for the room that can be used in room links.
        created_at:
          $ref: "#/components/schemas/Timestamp"
          description: The date when the room was created
        created_by:
          $ref: "#/components/schemas/PublicUserProfile"
          description: "The public user profile of the room's owner"
        guest_access:
          $ref: "#/components/schemas/GuestAccess"
          description: Guest access mode
        id:
          $ref: "#/components/schemas/RoomId"
          description: The ID of the room
        password:
          $ref: "#/components/schemas/RoomPassword"
          description: "The password of the room, if any"
        waiting_room:
          type: boolean
          description: If waiting room is enabled
      example:
        alias: personal-room-name_0000000000000000
        created_at: "1970-01-01T00:00:00Z"
        created_by:
          avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
          display_name: Alice Adams
          email: alice@example.com
          firstname: Alice
          id: 00000000-0000-0000-0000-0000000a11c3
          lastname: Adams
          title: ""
        guest_access: waiting_room
        id: 00000000-0000-0000-0000-000000000000
        password: v3rys3cr3t
        waiting_room: false
...
    SipConfigResource:
      type: object
      description: "Response for the `GET /rooms/{room_id}/sip` endpoint"
      required:
        - room
        - sip_id
        - password
        - lobby
      properties:
        lobby:
          type: boolean
          description: Flag if the room is a lobby
        password:
          $ref: "#/components/schemas/CallInPassword"
          description: The SIP password
        room:
          $ref: "#/components/schemas/RoomId"
          description: The room id
        sip_id:
          $ref: "#/components/schemas/CallInId"
          description: The SIP ID
      example:
        lobby: false
        password: "9876543210"
        room: 00000000-0000-0000-0000-0000abadcafe
        sip_id: "0123456789"
...
    Timestamp:
      type: string
      format: date-time
      description: |-
        A UTC DateTime wrapper that implements ToRedisArgs and FromRedisValue.

        The values are stores as unix timestamps in redis.
...
    UserId:
      type: string
      format: uuid
      description: The id of a user
      example: 00000009-9889-9889-9889-988000000000
...
  responses:
...
    Forbidden:
      description: The authorized user has no permission to access the requested resource
    InternalServerError:
      description: An internal server error occurred
    NotFound:
      description: The requested data could not be found
    Unauthorized:
      description: |-
        The provided access token is expired or the provided id or access token is invalid.
                The WWW-Authenticate header will contain an error description 'session expired' to distinguish between
                an invalid and an expired token
      headers:
        www-authenticate:
          schema:
            type: string
          description: |

            Will contain 'session expired' to distinguish between an invalid and an expired token.

            Examples:

                Bearer error="invalid_token", error_description="The provided access token is invalid"
                Bearer error="invalid_token", error_description="The user session expired"
      content:
        application/json:
          schema:
            type: object
            description: Internal reusable dummy type for utoipa unauthorized error
            required:
              - code
              - message
            properties:
              code:
                type: string
                description: Machine readable error code
              message:
                type: string
                description: Human readable message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    InviteCode:
      type: http
      scheme: bearer
...